@@ -25,18 +25,18 @@ class IconToggle extends StatefulWidget {
   final Color activeColor;
   final Color inactiveColor;
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
   final Duration duration;
-  final Duration reverseDuration;
+  final Duration? reverseDuration;
   @override
   _IconToggleState createState() => _IconToggleState();
 }
 
 class _IconToggleState extends State<IconToggle>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _position;
+  AnimationController? _controller;
+  Animation<double>? _position;
   bool _cancel = false;
 
   @override
@@ -46,12 +46,12 @@ class _IconToggleState extends State<IconToggle>
         vsync: this,
         duration: Duration(milliseconds: 100),
         reverseDuration: Duration(milliseconds: 50));
-    _position = CurvedAnimation(parent: _controller, curve: Curves.linear);
-    _position.addStatusListener((status) {
+    _position = CurvedAnimation(parent: _controller!, curve: Curves.linear);
+    _position!.addStatusListener((status) {
       if (status == AnimationStatus.dismissed &&
           widget.onChanged != null &&
           _cancel == false) {
-        widget.onChanged(!widget.value);
+        widget.onChanged!(!widget.value);
       }
     });
   }
@@ -80,7 +80,7 @@ class _IconToggleState extends State<IconToggle>
       child: Padding(
         padding: const EdgeInsets.all(10.0),
         child: _IconToggleable<double>(
-          listenable: _position,
+          listenable: _position!,
           activeColor: widget.activeColor,
           inactiveColor: widget.inactiveColor,
           child: AnimatedSwitcher(
@@ -102,21 +102,21 @@ class _IconToggleState extends State<IconToggle>
 
 class _IconToggleable<T> extends AnimatedWidget {
   _IconToggleable({
-    Animation<T> listenable,
+    Animation<T>? listenable,
     this.activeColor,
     this.inactiveColor,
     this.child,
-  }) : super(listenable: listenable);
-  final Color activeColor;
-  final Color inactiveColor;
-  final Widget child;
+  }) : super(listenable: listenable!);
+  final Color? activeColor;
+  final Color? inactiveColor;
+  final Widget? child;
   @override
   Widget build(BuildContext context) {
     return CustomPaint(
       painter: _IconPainter(
-        position: listenable,
-        activeColor: activeColor,
-        inactiveColor: inactiveColor,
+        position: listenable as Animation<double>,
+        activeColor: activeColor!,
+        inactiveColor: inactiveColor!,
       ),
       child: child,
     );
@@ -129,16 +129,16 @@ class _IconPainter extends CustomPainter {
     this.activeColor,
     this.inactiveColor,
   });
-  final Animation<double> position;
-  final Color activeColor;
-  final Color inactiveColor;
+  final Animation<double>? position;
+  final Color? activeColor;
+  final Color? inactiveColor;
 
-  double get _value => position != null ? position.value : 0;
+  double get _value => position != null ? position!.value : 0;
 
   @override
   void paint(Canvas canvas, Size size) {
     final Paint paint = Paint()
-      ..color = Color.lerp(inactiveColor, activeColor, _value)
+      ..color = Color.lerp(inactiveColor, activeColor, _value)!
           .withOpacity(math.min(_value, 0.15))
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.0;
