@@ -1,6 +1,6 @@
-import 'dart:async';
 import 'dart:io';
-import 'file:///Users/makisu/flutter-icons/test/a.dart';
+// import 'dart:async';
+// import 'file:///Users/makisu/flutter-icons/test/a.dart';
 //void main() async {
 //  File file = File("././lib/src/a.dart");
 //  if (!file.existsSync())
